+++ src/error.rs
@@ -1,5 +1,5 @@
-//! Additional error types designed to retain as much information as possible
-//! during a parse failure, making full use of the [`ParseError`] trait.
+//! An error type, [`ErrorTree`], designed to retain much more useful
+//! information about parse failures than the built-in nom error types.
 
 use std::{
     error::Error,
@@ -19,18 +19,24 @@ use crate::tag::TagError;
 
 /// Enum for generic things that can be expected by nom parsers
 ///
-/// Certain nom parsers (think digit1 or space1) are "low level" in the sense
-/// that they scan for a specific character or kind of character. This enum
-/// tracks the different kinds of things that can be expected and not found
-/// at a location.
+/// Certain nom parsers (think [`digit1`], [`tag`], or [`space1`]) are "base
+/// level" in the sense that, rather than combining subparsers, they scan for a
+/// specific character or specific kind of character. This enum tracks the
+/// different kinds of things that can be expected by these base parses.
 ///
-/// Printing an expectation via `Display` will only include the thing that
+/// Printing an expectation via [`Display`] will only include the thing that
 /// was expected, in a form suitable for being prefixed with "expected" or
 /// suffixed with "was expected".
 ///
 /// This enum is non-exhaustive; it is intended to represent everything parse
 /// errors where we know *specifically* what was expected. For instance,
-/// take_while cannot create an Expectation, because it can't meaningfully
+/// [`take_while`] cannot create an [`Expectation`], because it can't
+/// meaningfully report what its subparser is expecting.
+///
+/// [`digit1`]: nom::character::complete::digit1
+/// [`tag`]: crate::tag::complete::tag
+/// [`space1`]: nom::character::complete::space1
+/// [`take_while`]: nom::bytes::complete::take_while
 #[non_exhaustive]
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub enum Expectation {
@@ -95,14 +101,18 @@ impl Display for Expectation {
 /// [`ErrorTree`].
 #[derive(Debug)]
 pub enum BaseErrorKind {
-    /// Something specific was expected.
+    /// Something specific was expected, such as a specific
+    /// [character][Expectation::Char] or any [digit](Expectation::Digit).
+    /// See [`Expectation`] for details.
     Expected(Expectation),
 
     /// A nom parser failed.
     Kind(NomErrorKind),
 
     /// An error outside of nom occurred during parsing; for instance, as a
-    /// result of an error during `map_res`.
+    /// result of an error during [`map_res`].
+    ///
+    /// [`map_res`]: crate::parser_ext::ParserExt::map_res
     // Design note: I've gone back and forth on whether or not to exclude the
     // ErrorKind from this variant. Right now I'm doing so, because it seems
     // like in practice it's *always* MapRes.
@@ -113,23 +123,27 @@ impl Display for BaseErrorKind {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         match *self {
             BaseErrorKind::Expected(expectation) => write!(f, "expected {}", expectation),
-            BaseErrorKind::External(ref err) => write!(f, "external error: \"{}\"", err),
+            BaseErrorKind::External(ref err) => {
+                writeln!(f, "external error:")?;
+                let mut f = IndentWriter::new("  ", f);
+                write!(f, "{}", err)
+            }
             BaseErrorKind::Kind(kind) => write!(f, "error in {:?}", kind),
         }
     }
 }
 
-/// Context that can appear in a stack, above a base [`ErrorTree`]. Stack
-/// contexts are attached by parsers to errors from subparsers during stack
-/// unwinding.
-#[derive(Debug, Clone)]
+/// Context that can appear in a [stack][ErrorTree::Stack], above a base
+/// [`ErrorTree`]. Stack contexts are attached by parser combinators to errors
+/// from their subparsers during stack unwinding.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum StackContext {
     /// A nom combinator attached an [`ErrorKind`][NomErrorKind] as context
     /// for a subparser error.
     Kind(NomErrorKind),
 
-    /// The [`context`] combinator attached a message as context for a
-    /// subparser error.
+    /// The [`context`][crate::parser_ext::ParserExt::context] combinator
+    /// attached a message as context for a subparser error.
     Context(&'static str),
 }
 
@@ -139,23 +153,6 @@ impl Display for StackContext {
             StackContext::Kind(kind) => write!(f, "while parsing {:?}", kind),
             StackContext::Context(ctx) => write!(f, "in section {:?}", ctx),
         }
-    }
-}
-
-/// A generic struct combining a specific location with some kind of error
-/// information.
-#[derive(Debug, Clone)]
-pub struct AtLocation<I, T> {
-    /// The location associated with this info
-    pub location: I,
-
-    /// The info. Usually a [`BaseErrorKind`] or [`StackContext`].
-    pub info: T,
-}
-
-impl<I: Display, T: Display> Display for AtLocation<I, T> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "{} at {:#}", self.info, self.location)
     }
 }
 
@@ -173,19 +170,209 @@ impl<I: Display, T: Display> Display for AtLocation<I, T> {
 /// very little useful contextual information to error traces; for example,
 /// [`ErrorKind::Alt`] doesn't add any interesting context to an
 /// [`ErrorTree::Alt`], and its presence in a stack precludes merging together
-/// adjacent sets of [`ErrorTree::Alt`] siblings
+/// adjacent sets of [`ErrorTree::Alt`] siblings.
 ///
-/// [`VerboseError`]: nom::error::VerboseError
-/// [`ErrorKind`]: nom::error::ErrorKind
-/// [`ErrorKind::Alt`]: nom::error::ErrorKind::Alt
+/// # Formatting
+///
+/// The *exact* [`Display`]
+///
+/// # Examples
+///
+/// ## Base parser errors
+///
+/// An `ErrorTree::Base` is an error that occurred at the "bottom" of the stack,
+/// from a parser looking for 1 specific kind of thing.
+///
+/// ```rust
+/// use cool_asserts::assert_matches;
+/// use nom::{Parser, Err};
+/// use nom::character::complete::{digit1, char};
+/// use nom_supreme::error::{ErrorTree, BaseErrorKind, StackContext, Expectation};
+/// use nom_supreme::parser_ext::ParserExt;
+///
+/// let err: Err<ErrorTree<&str>> = digit1.parse("abc").unwrap_err();
+///
+/// assert_matches!(err, Err::Error(ErrorTree::Base{
+///     location: "abc",
+///     kind: BaseErrorKind::Expected(Expectation::Digit),
+/// }));
+///
+/// let err: Err<ErrorTree<&str>> = char('a').and(char('b')).parse("acb").unwrap_err();
+///
+/// assert_matches!(err, Err::Error(ErrorTree::Base{
+///     location: "cb",
+///     kind: BaseErrorKind::Expected(Expectation::Char('b')),
+/// }));
+/// ```
+///
+/// ## Stacks
+///
+/// An [`ErrorTree::Stack`] is created when a parser combinator—typically
+/// [`context`]—attaches additional error context to a subparser error. It can
+/// have any [`ErrorTree`] at the base of the stack.
+///
+/// ```rust
+/// use cool_asserts::assert_matches;
+/// use nom::{Parser, Err};
+/// use nom::character::complete::{alpha1, space1, char,};
+/// use nom::sequence::{separated_pair, delimited};
+/// use nom_supreme::parser_ext::ParserExt;
+/// use nom_supreme::error::{ErrorTree, BaseErrorKind, StackContext, Expectation};
+///
+/// // Parse a single identifier, defined as just a string of letters.
+/// let identifier = alpha1.context("identifier");
+///
+/// // Parse a pair of identifiers, separated by whitespace
+/// let identifier_pair = separated_pair(identifier, space1, identifier)
+///     .context("identifier pair");
+///
+/// // Parse a pair of identifiers in parenthesis.
+/// let mut parenthesized = delimited(char('('), identifier_pair, char(')'))
+///     .context("parenthesized");
+///
+/// let err: Err<ErrorTree<&str>> = parenthesized.parse("(abc 123)").unwrap_err();
+///
+/// assert_matches!(err, Err::Error(ErrorTree::Stack {
+///     base,
+///     contexts,
+/// }) => {
+///     assert_matches!(*base, ErrorTree::Base {
+///         location: "123)",
+///         kind: BaseErrorKind::Expected(Expectation::Alpha)
+///     });
+///
+///     assert_eq!(contexts, [
+///         ("123)", StackContext::Context("identifier")),
+///         ("abc 123)", StackContext::Context("identifier pair")),
+///         ("(abc 123)", StackContext::Context("parenthesized")),
+///     ]);
+/// });
+/// ```
+///
+/// ## Alternatives
+///
+/// An [`ErrorTree::Alt`] is created when a series of parsers are all tried,
+/// and all of them fail. Most commonly this will happen via the [`alt`]
+/// combinator or the equivalent [`.or`] postfix combinator. When all of these
+/// subparsers fail, their errors (each individually their own `ErrorTree`) are
+/// aggregated into an [`ErrorTree::Alt`], indicating that "any one of these
+/// things were expected."
+///
+/// ```rust
+/// use cool_asserts::assert_matches;
+/// use nom::{Parser, Err};
+/// use nom::branch::alt;
+/// use nom_supreme::error::{ErrorTree, BaseErrorKind, StackContext, Expectation};
+/// use nom_supreme::parser_ext::ParserExt;
+/// use nom_supreme::tag::complete::tag;
+///
+/// let parse_bool = alt((
+///     tag("true").value(true),
+///     tag("false").value(true),
+/// ));
+///
+/// let mut parse_null_bool = alt((
+///     parse_bool.map(Some),
+///     tag("null").value(None),
+/// ));
+///
+/// let err: Err<ErrorTree<&str>> = parse_null_bool.parse("123").unwrap_err();
+///
+/// // This error communicates to the caller that any one of "true", "false",
+/// // or "null" was expected at that location.
+/// assert_matches!(err, Err::Error(ErrorTree::Alt(choices)) => {
+///     assert_matches!(choices.as_slice(), [
+///         ErrorTree::Base {
+///             location: "123",
+///             kind: BaseErrorKind::Expected(Expectation::Tag("true"))},
+///         ErrorTree::Base {
+///             location: "123",
+///             kind: BaseErrorKind::Expected(Expectation::Tag("false"))},
+///         ErrorTree::Base {
+///             location: "123",
+///             kind: BaseErrorKind::Expected(Expectation::Tag("null"))},
+///     ])
+/// });
+/// ```
+///
+/// ## Contexts and Alternatives
+///
+/// Because [`Stack`] and [`Alt`] recursively contain [`ErrorTree`] errors from
+/// subparsers, they can be can combined to create error trees of arbitrary
+/// complexity.
+///
+/// ```rust
+/// use cool_asserts::assert_matches;
+/// use nom::{Parser, Err};
+/// use nom::branch::alt;
+/// use nom_supreme::error::{ErrorTree, BaseErrorKind, StackContext, Expectation};
+/// use nom_supreme::parser_ext::ParserExt;
+/// use nom_supreme::tag::complete::tag;
+///
+/// let parse_bool = alt((
+///     tag("true").value(true),
+///     tag("false").value(true),
+/// )).context("bool");
+///
+/// let mut parse_null_bool = alt((
+///     parse_bool.map(Some),
+///     tag("null").value(None).context("null"),
+/// )).context("null or bool");
+///
+/// let err: Err<ErrorTree<&str>> = parse_null_bool.parse("123").unwrap_err();
+///
+/// assert_matches!(err, Err::Error(ErrorTree::Stack{base, contexts}) => {
+///     assert_eq!(contexts, [("123", StackContext::Context("null or bool"))]);
+///     assert_matches!(*base, ErrorTree::Alt(choices) => {
+///         assert_matches!(&choices[0], ErrorTree::Stack{base, contexts} => {
+///             assert_eq!(contexts, &[("123", StackContext::Context("bool"))]);
+///             assert_matches!(&**base, ErrorTree::Alt(choices) => {
+///                 assert_matches!(&choices[0], ErrorTree::Base {
+///                     location: "123",
+///                     kind: BaseErrorKind::Expected(Expectation::Tag("true"))
+///                 });
+///                 assert_matches!(&choices[1], ErrorTree::Base {
+///                     location: "123",
+///                     kind: BaseErrorKind::Expected(Expectation::Tag("false"))
+///                 });
+///            });
+///         });
+///         assert_matches!(&choices[1], ErrorTree::Stack{base, contexts} => {
+///             assert_eq!(contexts, &[("123", StackContext::Context("null"))]);
+///             assert_matches!(&**base, ErrorTree::Base {
+///                 location: "123",
+///                 kind: BaseErrorKind::Expected(Expectation::Tag("null"))
+///             });
+///         });
+///     });
+/// });
+/// ```
+///
+/// # Display formatting
+///
+/// TODO WRITE THIS SECTION
+///
+/// [`.or`]: nom::Parser::or
+/// [`Alt`]: ErrorTree::Alt
 /// [`alt`]: nom::branch::alt
+/// [`context`]: nom::error::context
+/// [`ErrorKind::Alt`]: nom::error::ErrorKind::Alt
+/// [`ErrorKind`]: nom::error::ErrorKind
+/// [`Stack`]: ErrorTree::Stack
+/// [`VerboseError`]: nom::error::VerboseError
 #[derive(Debug)]
 pub enum ErrorTree<I> {
     /// A specific error event at a specific location. Often this will indicate
     /// that something like a tag or character was expected at that location.
     /// When used as part of a stack, it indicates some additional context for
     /// the root error of the stack.
-    Base(AtLocation<I, BaseErrorKind>),
+    Base {
+        /// The location of this error in the input
+        location: I,
+
+        /// The specific error that occurred
+        kind: BaseErrorKind,
+    },
 
     /// A stack indicates a chain of error contexts was provided. The stack
     /// should be read "backwards"; that is, errors *earlier* in the `Vec`
@@ -195,7 +382,7 @@ pub enum ErrorTree<I> {
         base: Box<Self>,
 
         /// The stack of contexts attached to that error
-        contexts: Vec<AtLocation<I, StackContext>>,
+        contexts: Vec<(I, StackContext)>,
     },
 
     /// A series of parsers were tried in order at the same location (for
@@ -219,18 +406,15 @@ impl<I> ErrorTree<I> {
         // only happens when alternating between different *kinds* of
         // ErrorTree; nested groups of Alt or Stack are flattened.
         match self {
-            ErrorTree::Base(AtLocation { location, info }) => ErrorTree::Base(AtLocation {
+            ErrorTree::Base { location, kind } => ErrorTree::Base {
                 location: convert_location(location),
-                info,
-            }),
+                kind,
+            },
             ErrorTree::Stack { base, contexts } => ErrorTree::Stack {
                 base: Box::new(base.map_locations_ref(convert_location)),
                 contexts: contexts
                     .into_iter()
-                    .map(|AtLocation { location, info }| AtLocation {
-                        location: convert_location(location),
-                        info,
-                    })
+                    .map(|(location, context)| (convert_location(location), context))
                     .collect(),
             },
             ErrorTree::Alt(siblings) => ErrorTree::Alt(
@@ -254,12 +438,11 @@ impl<I> ErrorTree<I> {
 impl<I: Display> Display for ErrorTree<I> {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         match self {
-            ErrorTree::Base(err) => err.fmt(f),
+            ErrorTree::Base { location, kind } => write!(f, "{} at {:#}", kind, location),
             ErrorTree::Stack { contexts, base } => {
-                contexts
-                    .iter()
-                    .rev()
-                    .try_for_each(|ctx| writeln!(f, "{},", ctx))?;
+                contexts.iter().rev().try_for_each(|(location, context)| {
+                    writeln!(f, "{} at {:#},", context, location)
+                })?;
 
                 base.fmt(f)
             }
@@ -304,10 +487,7 @@ impl<I: InputLength> ParseError<I> for ErrorTree<I> {
             kind => BaseErrorKind::Kind(kind),
         };
 
-        ErrorTree::Base(AtLocation {
-            location,
-            info: kind,
-        })
+        ErrorTree::Base { location, kind }
     }
 
     /// Combine an existing error with a new one. This is how error context is
@@ -317,10 +497,7 @@ impl<I: InputLength> ParseError<I> for ErrorTree<I> {
     /// If `other` is already an `ErrorTree::Stack`, the context is added to
     /// the stack; otherwise, a new stack is created, with `other` at the root.
     fn append(location: I, kind: NomErrorKind, other: Self) -> Self {
-        let context = AtLocation {
-            location,
-            info: StackContext::Kind(kind),
-        };
+        let context = (location, StackContext::Kind(kind));
 
         match other {
             // Don't create a stack of [ErrorKind::Alt, ErrorTree::Alt(..)]
@@ -345,10 +522,10 @@ impl<I: InputLength> ParseError<I> for ErrorTree<I> {
 
     /// Create an error indicating an expected character at a given position
     fn from_char(location: I, character: char) -> Self {
-        ErrorTree::Base(AtLocation {
+        ErrorTree::Base {
             location,
-            info: BaseErrorKind::Expected(Expectation::Char(character)),
-        })
+            kind: BaseErrorKind::Expected(Expectation::Char(character)),
+        }
     }
 
     /// Combine two errors from branches of alt. If either or both errors are
@@ -356,6 +533,8 @@ impl<I: InputLength> ParseError<I> for ErrorTree<I> {
     /// otherwise, a new [`ErrorTree::Alt`] is created, containing both
     /// `self` and `other`.
     fn or(self, other: Self) -> Self {
+        // For now we assume that there's no need to try and preserve
+        // left-to-right ordering of alternatives.
         let siblings = match (self, other) {
             (ErrorTree::Alt(siblings1), ErrorTree::Alt(siblings2)) => {
                 match siblings1.capacity() >= siblings2.capacity() {
@@ -377,10 +556,7 @@ impl<I: InputLength> ParseError<I> for ErrorTree<I> {
 impl<I> ContextError<I> for ErrorTree<I> {
     /// Similar to append: Create a new error with some added context
     fn add_context(location: I, ctx: &'static str, other: Self) -> Self {
-        let context = AtLocation {
-            location,
-            info: StackContext::Context(ctx),
-        };
+        let context = (location, StackContext::Context(ctx));
 
         match other {
             // This is already a stack, so push on to it
@@ -404,19 +580,19 @@ impl<I> ContextError<I> for ErrorTree<I> {
 impl<I, E: Error + Send + Sync + 'static> FromExternalError<I, E> for ErrorTree<I> {
     /// Create an error from a given external error, such as from FromStr
     fn from_external_error(location: I, _kind: NomErrorKind, e: E) -> Self {
-        ErrorTree::Base(AtLocation {
+        ErrorTree::Base {
             location,
-            info: BaseErrorKind::External(Box::new(e)),
-        })
+            kind: BaseErrorKind::External(Box::new(e)),
+        }
     }
 }
 
 impl<I> TagError<I, &'static str> for ErrorTree<I> {
     fn from_tag(location: I, tag: &'static str) -> Self {
-        ErrorTree::Base(AtLocation {
+        ErrorTree::Base {
             location,
-            info: BaseErrorKind::Expected(Expectation::Tag(tag)),
-        })
+            kind: BaseErrorKind::Expected(Expectation::Tag(tag)),
+        }
     }
 }